users/urls.py
from django.urls import path
from .views import (
    RegisterView, PasswordResetBySecretView, ProfileView
)
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView

app_name = 'users'

urlpatterns = [
    path('register/', RegisterView.as_view(), name='register'),
    path('login/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('profile/', ProfileView.as_view(), name='profile'),
    path('password-reset-secret/', PasswordResetBySecretView.as_view(), name='password_reset_secret'),
]